add vwadd.vv, vwsubu.vv, vwsub.vv
fix store instructions when SEW != EEW

+++ src/instructions/v_register.rs
@@ -19,9 +19,9 @@ impl Default for VRegister {
 }
 
 impl VRegister {
-    pub fn clone_with_eew(&self, veew: u32) -> VRegister {
+    pub fn clone_with_eew(&self, veew: usize) -> VRegister {
         let bytes = self.to_le_bytes();
-        VRegister::from_le_bytes(veew, bytes)
+        VRegister::from_le_bytes(veew as u32, bytes)
     }
 
     pub fn to_le_bytes(&self) -> [u8; 256] {
